Move helpers from Report module to separate module
The Report module containing the base class for all reports contained
multiple functions that aren’t report-specific. Move these to a separate
new module for helper functions.
